DebugBallSpawner progress
When added to a scene, it will get all of the child collision
objects on specified layers and register to handle their input
event signals. When a left click is received, the global coordinates
of that click are stored, and when the click is released, those
coordinates are stored as well.