figured out what I was doing wrong RE debug arrow
I made the debug arrow node top-level and set its transform
to identity when adding it to the rollover. From there I
multiplied its basis by the global transform basis and  it
lined up as I would expect. Problem solved and I've confirmed
that the force is now aligned properly.